Add global variable scope
This works by making a single shared scope across all files. When you
open or require a file that has globals in it, they get added, but we
don't crawl your whole project directory for all globals yet. We are
still ignorant of _ENV or setfenv, so there _will_ be false positives in
those cases. Also, if you delete a global from a file, it won't actually
be removed from the global scope, which is something I hope to fix
sometime soon.