Add smarter return support
We now track multireturns, which is like a "yeah dummy" kind of feature
to be missing, and we do a better job of handling multiple return sites
with the same signature, which seems to happen for reasons.
I'm not really sure what past me had in mind when they created the
"Literal" tag but I think this is the right way to handle it in the
function sig code.